+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
  - Seek ability to player controls
 
+### Fix
+ - Add parent param to player popout
+
 ## 1.1.3 - 2020-05-02
 ### Added
  - Chat history loader feature