+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Fix
  - Add parent param to player popout
+ - Update player controls
 
 ## 1.1.3 - 2020-05-02
 ### Added